backend/pyproject.toml
[tool.poetry]
name = "backend-server"
version = "0.1.0"
description = ""
authors = ["Adel, Ivan <a4got10king@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
asgiref = "^3.8.1"
django = "^5.0.3"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
sqlparse = "^0.4.4"
typing-extensions = "^4.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
